moved acceptance test to routing tests
